+++ src/Cart/resources/database/migrations/2023_01_12_123641_add_cart_item_configuration.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+return new class () extends Migration {
+    public function up()
+    {
+        Schema::table('cart_items', function (Blueprint $table) {
+            $table->json('configuration')->nullable()->after('price');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('cart_items', function (Blueprint $table) {
+            $table->dropColumn('configuration');
+        });
+    }
+};